+++ crates/ironcalc-base/src/expressions/parser/static_analysis.rs
@@ -911,6 +911,13 @@ fn get_function_args_signature(kind: &Function, arg_count: usize) -> Vec<Signatu
             }
             result
         }
+        Function::Filter => {
+            let mut result = vec![Signature::Vector; arg_count];
+            for item in result.iter_mut().skip(2) {
+                *item = Signature::Scalar;
+            }
+            result
+        }
         Function::Aggregate => {
             let mut result = vec![Signature::Vector; arg_count];
             if arg_count >= 1 {
@@ -1503,6 +1510,7 @@ fn static_analysis_on_function(kind: &Function, args: &[Node]) -> StaticResult {
         Function::Sequence => not_implemented(args),
         Function::Sort => not_implemented(args),
         Function::Unique => not_implemented(args),
+        Function::Filter => not_implemented(args),
         Function::Aggregate => not_implemented(args),
         Function::Rand => not_implemented(args),
         Function::Randbetween => scalar_arguments(args),